+++ backend/src/main/java/com/whereismyhome/amenities/entity/Cctv.java
@@ -6,7 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Entity(name = "cctv")
+@Entity
+@Table(name = "cctv")
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
