+++ backend/src/main/java/com/whereismyhome/amenities/service/AmenService.java
@@ -1,7 +1,9 @@
 package com.whereismyhome.amenities.service;
 
 import com.whereismyhome.amenities.dto.response.BusResponseDto;
+import com.whereismyhome.amenities.dto.response.CctvResponseDto;
 import com.whereismyhome.amenities.repository.BusRepository;
+import com.whereismyhome.amenities.repository.CctvRepository;
 import com.whereismyhome.houseinfo.dto.HouseResponseDto;
 import com.whereismyhome.houseinfo.repository.HouseInfoRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +16,7 @@ import java.util.List;
 public class AmenService {
     private final HouseInfoRepository houseInfoRepository;
     private final BusRepository busRepository;
+    private final CctvRepository cctvRepository;
 
     //반경 내 아파트 정보 조회
     public List<HouseResponseDto> findApt(String lng, String lat) {
@@ -27,5 +30,12 @@ public class AmenService {
         List<BusResponseDto> busList = busRepository.findByBus(lng, lat);
 
         return busList;
+    }
+
+    //변경 내 cctv 정보조회
+    public List<CctvResponseDto> findCctv(String lng, String lat) {
+        List<CctvResponseDto> cctvList = cctvRepository.findByCctv(lng, lat);
+
+        return cctvList;
     }
 }
